refactor(MoveResults): MoveResults, MoveResultsPerLine 패키지를 domain으로 이동
패키지를 위와 같이 이동하였습니다. 이유는 BridgeGame에서 사용하기 때문에, 같은 패키지 내에 두어야 된다고 판단했습니다.
